refactor(#31): update NotEmptyCollection.size property

+++ src/commonMain/kotlin/kotools/types/collection/NotEmptyCollection.kt
@@ -16,8 +16,7 @@ public sealed interface NotEmptyCollection<out E> {
 
     /** The size of this collection. */
     public val size: StrictlyPositiveInt
-        get() = tail.sizeOrZero.toInt()
-            .plus(1)
+        get() = tail.sizeOrZero.plus(1)
             .toStrictlyPositiveInt()
             .getOrThrow()
